pyproject.toml
[project]
name = "pikvm_lib"
version = "0.4.4"
authors = [
  { name="guanana", email="guanana2@gmail.com" },
]
description = "Library to control PiKVM with Python"
readme = "README.md"
keywords = ["pikvm", "pikvm-lib", "kvm"]
license = {text = "GNU v3"}
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pyotp<=2.9.0",
    "websocket-client>=1.7.0",
    "pillow>=11.3.0"
]


classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/guanana/pikvm-lib"
Issues = "https://github.com/guanana/pikvm-lib/issues"

[tool.bumpver]
current_version = "0.4.4"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = true
tag = true
push = false

[tool.setuptools.packages.find]
where = ["src"]

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
"README.md" = [
    "{version}",
    "{pep440_version}",
]
"src/pikvm_lib/__init__.py" = ["{version}"]
